+++ backend/src/main/java/com/optimize/elykia/core/entity/report/CashDeposit.java
@@ -37,6 +37,9 @@ public class CashDeposit extends Auditable<String> {
     @Column(columnDefinition = "double precision default 0")
     private Double newBalanceAmount = 0.0;
 
+    @Column(columnDefinition = "double precision default 0")
+    private Double surplusAmount = 0.0;
+
     @Column(columnDefinition = "TEXT")
     private String billetage; // JSON storing { "10000": 5, "5000": 2 }